refactor: rename some clases
